Add override for clock_gettime() in libcompat for systems that do not supply it
clock_gettime() is defined in POSIX.1-2001. However, some systems
(notably OSX) do not provide it. To allow such system to compile check,
clock_gettime() is added to libcompat. In the libcompat version, equivalent
calls for OSX are added to get time. For other systems, clock_gettime()
simply sets the time to 0.