fix compiler warning: subunit header file not found
The code is using ENABLE_SUBUNIT to determine if subunit is available.
HAVE_SUBUNIT_CHILD_H is never defined. When subunit is used, its
header files are never found, causing a compiler warning. This commit
replaces the #if protecting the header to be consistent with the
rest of the code.