modify test_output.sh to compare output based on HAVE_FORK
The output of ex_output changes based on HAVE_FORK, as a test is
skipped when fork is unavailable. However, simply adding another
string mentioning one less test is not enough to get this working
for Windows...

When run in Windows, the output differs slightly from when it is run
under *NIX due to line endings. *NIX uses \n, whereas Windows
uses \r\n. To account for this, printf is used to generate the
Windows expected string.