Do not expect the duration for no-fork mode tests to be negative
For tests using fork, if the test fails exit() is called, which results
in the duration not being recorded for the test.

no-fork mode instead requires all tests to return, meaning their
duration is always recorded.

The test_xml_output.sh test expected that all failed tests have a negative
duration, which is not the case for no-fork mode. Removing
this expectation, and instead forcing all durations to be >=0 for
no-fork mode.