Print an error if nothing is read from punpack()
There is some issue with MinGW-w64 builds run on wine, if multiple
unit test programs are run concurrently. Sometimes they will fail
to get something from punpack. It may be a failure to read the file,
or the file, or some other issue.

When this happens, later in receive_test_result the rmesg is free'd,
but only by freeing its elements first, which causes a segfault.

Instead of crashing, check should error out gracefully.